# Optimize serialization of byte buffers in Python SDK