@amacneil 2025-02-14
Is there a script for running performance tests like that? Would be nice to keep in the repo if its not already.