@bryfox 2025-02-14
> Is there a script for running performance tests like that? Would be nice to keep in the repo if its not already.

The point cloud generation was very close to the one included here, but used more points. I wrote the original scripts with cProfile; I'm not sure how much value there is in including that. Maybe worth adding benchmarks more formally, though?