@amacneil 2025-02-14
Sounds good I added scope to FG-9514